requirements.txt
azure-data-tables==12.4.2
azure-core==1.26.3
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
